requirements.txt
uvicorn
fastapi
sqlalchemy
pydantic
starlette
google-genai
python-dotenv
torch
chatterbox-tts
openai
ollama
elevenlabs